python/otlp/ext+ddtrace-instrument/requirements.txt
attrs==23.2.0
bytecode==0.15.1
cattrs==23.2.3
ddsketch==3.0.1
ddtrace==2.8.3
Deprecated==1.2.14
envier==0.5.1
exceptiongroup==1.2.1
importlib-metadata==7.0.0
opentelemetry-api==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-aws-lambda==0.45b0
opentelemetry-propagator-aws-xray==1.0.1
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
protobuf==5.26.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
wrapt==1.16.0
xmltodict==0.13.0
zipp==3.18.1
